usecase パッケージをリファクタリング_album
